separate parser from its execution
